refactor(/model): domain 패키지의 이름을 model로 변경
