+++ src/main/java/roomescape/reservation/model/ReservationDate.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.model;
 
 import java.time.LocalDate;
